+++ sway-core/src/concurrent_slab.rs
@@ -1,11 +1,6 @@
 use std::{fmt, sync::RwLock};
 
-use crate::{
-    declaration_engine::{declaration_id::DeclId, declaration_wrapper::DeclWrapper},
-    engine_threading::*,
-    type_system::TypeId,
-    TypeInfo,
-};
+use crate::{declaration_engine::*, engine_threading::*, type_system::TypeId, TypeInfo};
 
 #[derive(Debug)]
 pub(crate) struct ConcurrentSlab<T> {
